requirements.txt
absl-py==2.3.1
aioice==0.10.1
aiortc==1.13.0
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
av==14.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dnspython==2.7.0
flatbuffers==25.2.10
fonttools==4.59.0
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
ifaddr==0.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
mediapipe==0.10.21
ml_dtypes==0.5.3
narwhals==2.0.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
streamlit==1.47.1
streamlit-webrtc==0.63.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
